Moved NoteItem component to Components directory
