+++ src/components/NoteItem/NoteItem.tsx
